ApiStatusBadge is dupplicated in CertificateList and CertificateDetailsM
